refactor(tmux): move PromptKind to command/copymode (pre-keybindings-removal)
